Added .d
.d performs multiple replaces.
It takes a string such as "azby" which will replace all 'a' with 'z'
and all 'b' with 'y'.

If they are more than one arguments. Than this is essentially like
multiple replaces. So: `ab,cd,de,fg` so all `ab` would be replaced with
`cd` and all `de` would be replaced with `fg`