Major updates to auto-functions
Added functions:
- id(x) - x is defined
- functify(x,y) - Creates a function with x, somewhat dependant on if y
is defined. Examples: functify("+") => function(a,b){return a+b},
functify("!-") => function(a,b){return b-a}, functify("a") =>
function(a,b){return a.a()}, functify("a",1) => function(a,b){return
a.a(b)}

Improvements:
- The auto-functions now transpile to a string that can be fed into
functify.
- Added function S.å - functionally equivalent to S.q().å.
- A.ã, A.ä, A.å, S.ã, S.ä, and S.å now accept a second parameter, which
will be the start of the cum-reduce or the -1th element in the pair-off.
- N.m and N.w now accept multiple parameters.