Fixed Number.k() bug
When the result of Number.k() was a one-item array, that one item was a Number object instead of an ordinary number. This breaks code that checks for an item's existence in the array. For example, "2k ø2", which checks if the number 2 exists in the array of factors of 2, returns false when it should return true.

Example in Chrome dev console: http://i.imgur.com/4NawhOk.png

This was fixed similarly to how I see you handling some other Number functions: by casting the Number object to an ordinary number by using "+this".